# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "\u8ba2\u5355\u7cfb\u7edf",
        "version": "v0.0.1"
    },
    "basePath": "/lsx/public/index.php/v1",
    "paths": {
...
        "/lsx/public/index.php/v1/food/add": {
            "post": {
                "tags": [
                    "\u83dc\u54c1\u6a21\u5757"
                ],
                "summary": "\u65b0\u589e/\u66f4\u65b0\u83dc\u54c1\u63a5\u53e3",
                "description": "\u65b0\u589e/\u66f4\u65b0\u83dc\u54c1",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "foodId",
                        "in": "formData",
                        "description": "\u83dc\u54c1id\uff0c\u4e0d\u4f20\u4e3a\u65b0\u589e\uff0c\u4f20\u4e86\u4e3a\u66f4\u65b0",
                        "type": "string"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "\u83dc\u540d",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "img",
                        "in": "formData",
                        "description": "\u56fe\u7247",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "formData",
                        "description": "\u7c7b\u522b",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "description": "\u4ef7\u683c",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "offprice",
                        "in": "formData",
                        "description": "\u6298\u6263\u4ef7",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "status",
                        "in": "formData",
                        "description": "\u72b6\u6001",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u8bf7\u6c42\u6210\u529f"
                    }
                }
            }
        },
        "/lsx/public/index.php/v1/food/del": {
            "post": {
                "tags": [
                    "\u83dc\u54c1\u6a21\u5757"
                ],
                "summary": "\u5220\u9664\u83dc\u54c1",
                "description": "\u5220\u9664\u83dc\u54c1",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "foodId",
                        "in": "formData",
                        "description": "\u5220\u9664\u83dc\u54c1",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u8bf7\u6c42\u6210\u529f"
                    }
                }
            }
        },
        "/lsx/public/index.php/v1/food/status": {
            "post": {
                "tags": [
                    "\u83dc\u54c1\u6a21\u5757"
                ],
                "summary": "\u4fee\u6539\u83dc\u54c1\u72b6\u6001",
                "description": "\u4fee\u6539\u83dc\u54c1\u72b6\u6001",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "foodId",
                        "in": "formData",
                        "description": "\u83dc\u54c1id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "status",
                        "in": "formData",
                        "description": "\u72b6\u6001onshelf/offshelf",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u8bf7\u6c42\u6210\u529f"
                    }
                }
            }
        },
        "/lsx/public/index.php/v1/food/search": {
            "get": {
                "tags": [
                    "\u83dc\u54c1\u6a21\u5757"
                ],
                "summary": "\u83b7\u53d6\u83dc\u54c1\u5217\u8868",
                "description": "\u83b7\u53d6\u83dc\u54c1\u5217\u8868\uff0c\u53ef\u4ee5\u6839\u636etitle\u6216type\u6a21\u7cca\u67e5\u8be2",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "\u83b7\u53d6\u7b2c\u51e0\u9875\u6570\u636e",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "\u6bcf\u9875\u591a\u5c11\u6761\u6570\u636e",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "\u83dc\u54c1\u540d\u79f0",
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "\u83dc\u54c1\u7c7b\u578b",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u8bf7\u6c42\u6210\u529f"
                    }
                }
            }
        },
        "/lsx/public/index.php/v1/food/list": {
            "get": {
                "tags": [
                    "\u83dc\u54c1\u6a21\u5757"
                ],
                "summary": "\u83b7\u53d6\u5df2\u4e0a\u67b6\u83dc\u54c1\u5217\u8868",
                "description": "\u83b7\u53d6\u5df2\u4e0a\u67b6\u83dc\u54c1\u5217\u8868(status==1)",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "\u83b7\u53d6\u7b2c\u51e0\u9875\u6570\u636e",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "\u6bcf\u9875\u591a\u5c11\u6761\u6570\u636e",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u8bf7\u6c42\u6210\u529f"
                    }
                }
            }
        },
...
    },
...
}